Showcase Of The First Website I Made For A University Project
This is the first website I made as part of the coursework during first
year of university

It consists of a main page containing a Literature Review made as part
of the unit coursework, a discovery log of the making of the website and
a video that includes subtitles.

The website uses HTML5 and CSS which provides a responsive design, and
is made so it will work on all mayor browsers.

If you want to view it you just need to clone this repo, navigate to
this commit and start a http server. There is nothing else to do as this
site is entirely made up of plain html and css.